// User/src/enums/messages.enum.ts
export enum UserMessages {
  USER_NOT_FOUND = "User not found",
  CREATE_USER_FAILED = "Create users failed.",
  UPDATE_USER_FAILED = "Update user failed.",
  DELETE_USER_FAILED = "Delete user failed.",
  GET_USERS_FAILED = "Get users failed.",
  GET_USER_FAILED = "Get user failed.",
  USER_CREATED = "Users created successfully.",
  USER_UPDATED = "User updated successfully.",
  USER_DELETED = "User deleted successfully.",
}